refactor: CartTotal 컴포넌트 내의 로직을 useCartTotal 커스텀 훅으로 분리
